Splitter opp beregningsgrunnlag for barnepensjon og omstillingsstønad
